refactor(cmd): extract completion engine (#526)
